Split JSONRemoteHelper into a seperate file
